Move schema-to-port handling to a helper
