+++ src/ApiComponent.js
@@ -1,51 +1,6 @@
 const { Component } = require('noflo');
 const slug = require('slug');
-
-function parameterToNoFlo(param) {
-  let datatype;
-  switch (param.type) {
-    case 'string': {
-      datatype = 'string';
-      break;
-    }
-    case 'number': {
-      datatype = 'number';
-      break;
-    }
-    case 'integer': {
-      datatype = 'int';
-      break;
-    }
-    case 'boolean': {
-      datatype = 'boolean';
-      break;
-    }
-    case 'object': {
-      datatype = 'object';
-      break;
-    }
-    case 'array': {
-      datatype = 'array';
-      break;
-    }
-    default: {
-      datatype = 'all';
-    }
-  }
-  const nofloDef = {
-    datatype,
-    description: param.description,
-  };
-  if (param.enum) {
-    nofloDef.values = param.enum;
-  }
-
-  // Special cases
-  if (param.type === 'string' && param.format === 'date-time') {
-    nofloDef.datatype = 'date';
-  }
-  return nofloDef;
-}
+const { parameterToNoFlo } = require('./schemaHelpers');
 
 function createBasicComponent(method, definition) {
   const c = new Component();
